Add files via upload
The Code is based on python, I use Jupyter Notebook to show the ideas of solving 1D pendulum by 4th RK Method, and all the code is finished by myself.
The author is a college student from China major in Physics, so the discription is based on Chinese, sorry for your understanding. However, the codes there still work well.
It would be a great pleasure for coummunicate the pendulum case here, and I only develop the fundamental 1D pendulum and I am trying to solve the 3D part based on Langrane Equtaion.
If there is any question for you to understand the program, please contant Courage_bear@outlook.com
Thanks!